Progression tracking and question flow.
Progression tracking including first set of unit tests (sadly...will
need to do a lot of backfilling).
Front-end flow of correct/incorrect question handling.

+++ curricula/admin.py
@@ -112,10 +112,10 @@ class VectorAnswerForm(SpecialAnswerFormMixin, forms.ModelForm):
         model = Answer
         fields = ['magnitude', 'angle', 'x_component', 'y_component', 'is_correct', 'position']
 
-    magnitude = forms.IntegerField(required=False)
-    angle = forms.IntegerField(required=False)
-    x_component = forms.IntegerField(required=False)
-    y_component = forms.IntegerField(required=False)
+    magnitude = forms.FloatField(required=False)
+    angle = forms.FloatField(required=False)
+    x_component = forms.FloatField(required=False)
+    y_component = forms.FloatField(required=False)
 
 
 class VectorAnswerInline(NestedTabularInline):
